# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Tangula API接口文档\n唐古拉山脉，长江正源，滋润万物。\nTangula 寓意数据的源头，为研发团队提供即取即用的数据服务，帮助团队高效产出。",
        "title": "Tangula",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "zhang.jiejun@outlook.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8000",
    "paths": {
...
        "/store_pool/replica": {
            "get": {
                "description": "【登录权限】获取所有副本信息，普通用户与系统管理员只能获取自身创建的副本信息，超级管理员可以获取平台上所有的副本信息。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "副本管理"
                ],
                "summary": "获取所有副本信息",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "分页索引",
                        "name": "index",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 15,
                        "description": "分页数量",
                        "name": "count",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "副本状态过滤参数[0.不过滤，1024.未挂载，2048.挂载中，4096.已挂载",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "镜像类型过滤参数[0.不过滤类型，1001.FILE，1002.ISO，1003.CUSTOME，1004.VMWAREVM，1005.CASVM，1006.HCSVM，1007.OPENSTACKVM]",
                        "name": "dataType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"\"",
                        "description": "名称搜索过滤参数",
                        "name": "filter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "response": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/api.Paging"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "data": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/replica.Replica"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "【登录权限】创建一个新副本。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "副本管理"
                ],
                "summary": "创建副本",
                "parameters": [
                    {
                        "type": "string",
                        "description": "副本名称",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "副本类型[1001.File, 1002.ISO, 1003.CUSTOME, 1004.VMware, 1005.Cas, 1006.HCS, 1007.OpenStack]",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "\"\"",
                        "description": "描述",
                        "name": "desc",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "副本大小[单位：GB]",
                        "name": "size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "所在存储池uuid",
                        "name": "storePoolId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "response": {
                                            "$ref": "#/definitions/replica.Replica"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/store_pool/replica/:id/mount/instance": {
            "get": {
                "description": "【登录权限】获取副本执行记录。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "副本管理"
                ],
                "summary": "获取副本执行实例记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "目标副本ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "分页索引",
                        "name": "index",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 15,
                        "description": "分页数量",
                        "name": "count",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "response": {
                                            "$ref": "#/definitions/instance.Instance"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/store_pool/replica/batch/mount": {
            "post": {
                "description": "【登录权限】批量挂载副本到目标平台。\n\n【emphasis】 参数AppConfig需要根据应用区分，组织不同形式的[key:value]字符串。\n----Linux平台：\n--------【可选参数:string】mountPoint：目标挂载位置\n----VMware平台：\n--------【必填参数:string】locationPath：挂载清单位置\n--------【必填参数:string】computeResource：挂载计算资源\n--------【必填参数:bool】isRegisterVM：是否注册虚拟机\n--------【选填参数:string】vmName：指定虚拟机注册名称\n--------【选填参数:bool】powerOn：虚拟机注册后是否打开电源\n--------【选填参数:string】hostName：指定虚拟机操作系统主机名\n--------【选填参数:string】addr：指定虚拟机操作系统IP地址\n--------【选填参数:string】netMask：指定虚拟机操作系统网关\n--------【选填参数:string】gateWay：指定虚拟机操作系统掩码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "副本管理"
                ],
                "summary": "批量挂载副本",
                "parameters": [
                    {
                        "description": "批量挂载参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/view.BatchMountParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "response": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/replica.Replica"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/store_pool/replica/batch/unmount": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "副本管理"
                ],
                "parameters": [
                    {
                        "description": "批量挂载参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/view.BatchUnmountParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "response": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/replica.Replica"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/store_pool/replica/mount/instance/:id/logs": {
            "get": {
                "description": "【登录权限】获取执行记录详情。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "副本管理"
                ],
                "summary": "获取指定副本执行实例详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "目标实例ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "response": {
                                            "$ref": "#/definitions/instance.InstanceLog"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Paging": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "totalNum": {
                    "type": "integer"
                }
            }
        },
        "api.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "response": {
                    "type": "object"
                }
            }
        },
...
        "instance.Instance": {
            "type": "object",
            "properties": {
                "created_time": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mount_point": {
                    "type": "string"
                },
                "replica_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "target_id": {
                    "type": "integer"
                },
                "target_type": {
                    "type": "integer"
                },
                "type": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    }
}